Re-assign TRRS pad labels
This is a hacked symbol anyway, may as well make it a logical one. Now
the pad order flows from one end to another of the internal
connections (note that sleeve connection is at opposite end to pin so
*appears* out-of-order on PCB footprint). Also explicitly include pads
that are shorted together by reversible footprint and reflect this in
schematic.